+++ src/components/repository/RepositoryViewBase/RepositoryViewBase.vue
@@ -176,7 +176,7 @@ import RepositoryInfo from '@/components/repository/RepositoryInfo';
 import RepositoryNavigation from './RepositoryNavigation';
 import TrainModal from '@/components-v1/repository/TrainModal';
 import TrainResponse from '@/components-v1/repository/TrainResponse';
-import RequestAuthorizationForm from '@/components-v1/repository/RequestAuthorizationForm';
+import RequestAuthorizationForm from '@/components/repository/RequestAuthorizationForm';
 
 
 const ERROR_VERBOSE_LOOKUP = {
